fixing flaky tests
